🎨공통컴포넌트 분리 및 레이아웃 추가
