+++ src/components/index.js
@@ -1,0 +1,3 @@
+export { default as Footer } from "./Footer";
+export { default as Header } from "./Header";
+export { default as Nav } from "./Nav";
