+++ src/pages/Login.js
@@ -1,4 +1,4 @@
-const component = `
+const pageTemplate = `
     <main class="bg-gray-100 flex items-center justify-center min-h-screen">
       <div class="bg-white p-8 rounded-lg shadow-md w-full max-w-md">
         <h1 class="text-2xl font-bold text-center text-blue-600 mb-8">항해플러스</h1>
@@ -22,7 +22,7 @@ const component = `
     </main>
 `;
 
-function event(navigateTo) {
+function setupPage(onNavigateTo) {
   const form = document.querySelector("#login-form");
 
   form.addEventListener("submit", function (event) {
@@ -33,9 +33,9 @@ function event(navigateTo) {
         "user",
         JSON.stringify({ username, email: "", bio: "" })
       );
-      navigateTo("/");
+      onNavigateTo("/");
     }
   });
 }
 
-export default { component, event };
+export default { pageTemplate, setupPage };
