+++ src/pages/NotFound.js
@@ -1,5 +1,5 @@
-const component = `
-      <main class="bg-gray-100 flex items-center justify-center min-h-screen">
+const pageTemplate = `
+    <main class="bg-gray-100 flex items-center justify-center min-h-screen">
       <div class="bg-white p-8 rounded-lg shadow-md w-full text-center" style="max-width: 480px">
         <h1 class="text-2xl font-bold text-blue-600 mb-4">항해플러스</h1>
         <p class="text-4xl font-bold text-gray-800 mb-4">404</p>
@@ -14,4 +14,4 @@ const component = `
     </main>
 `;
 
-export default { component };
+export default { pageTemplate };
