+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export { default as Home } from "./Home";
+export { default as Login } from "./Login";
+export { default as NotFound } from "./NotFound";
+export { default as Profile } from "./Profile";
